block: scheduler: add sioplus scheduler
The Simple I/O scheduler is an extremely simple scheduler, based on noop and deadline, that relies on deadlines to ensure fairness. The algorithm does not do any sorting but basic merging, trying to keep a minimum overhead. It is aimed mainly for aleatory access devices (eg: flash devices). The plus version fixes writes_starved not being initialized on startup and also modifies the write starvation counting logic.